requirements.txt
discord.py>=2.3.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
colorlog>=6.7.0
greenlet>=3.0.0
